requirements.txt
bcrypt==4.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask==2.2.3
passlib==1.7.4
psycopg2==2.9.5
python-dotenv==1.0.0
python-miio==0.5.12
SQLAlchemy==2.0.7
WTForms==3.0.1
